refactor(core): move util functions into services
