+++ server/services/index.ts
@@ -1,7 +1,5 @@
-import fuzzySearchService from './fuzzySearchService';
 import settingsService from './settingsService';
 
 export default {
-  settingsService,
-  fuzzySearchService,
+  settingsService: settingsService,
 };
