--- server/utils/parsePagination.ts
@@ -1,14 +1,0 @@
-import { PaginationBaseQuery, paginationSchema } from '../config/querySchema';
-
-export const parsePagination = (paginationQuery: PaginationBaseQuery) => {
-  const {
-    page: pageQuery = '1',
-    pageSize: pageSizeQuery = '25',
-    withCount: withCountQuery = 'true',
-  } = paginationQuery;
-  const page = pageQuery ? parseInt(pageQuery, 10) : 1;
-  const pageSize = pageSizeQuery ? parseInt(pageSizeQuery, 10) : 25;
-  const withCount = withCountQuery ? withCountQuery === 'true' : true;
-
-  return { page, pageSize, withCount };
-};
